Rearrange content components and intercept article links (#30)
